Done following changes: 1) added CRUD apis for product 2) removed unused PersonController.java and its related classes 3) done code formatting of the whole projectPlease enter the commit message for your changes.

+++ src/main/java/com/dummy/service/DbInit.java
@@ -2,17 +2,18 @@ package com.dummy.service;
 
 import com.dummy.model.Product;
 import com.dummy.model.ProductDetailsResponse;
+import com.dummy.model.Quote;
+import com.dummy.model.QuoteDetailsResponse;
 import com.dummy.persistense.entity.ProductHE;
 import com.dummy.persistense.entity.QuotesHE;
 import com.dummy.persistense.repository.ProductRepository;
 import com.dummy.persistense.repository.QuotesRepository;
 import com.dummy.utils.Mapper;
-import com.dummy.model.Quote;
-import com.dummy.model.QuoteDetailsResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
+import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
@@ -40,29 +41,36 @@ public class DbInit {
         saveProducts(restTemplate);
     }
 
+    @Transactional
     private void saveQuotes(RestTemplate restTemplate) {
-        ResponseEntity<QuoteDetailsResponse> response = restTemplate
-                .getForEntity(GET_QUOTES, QuoteDetailsResponse.class);
-        QuoteDetailsResponse body = response.getBody();
-        List<Quote> quotes = body.getQuotes();
-        List<QuotesHE> list = quotes.stream().map(mapper::modelToEntity)
-                .collect(Collectors.toList());
-        quotesRepository.saveAll(list);
+        int size = quotesRepository.findAll().size();
+        if (size < 1 || size < 100) {
+            quotesRepository.deleteAll();
+            ResponseEntity<QuoteDetailsResponse> response = restTemplate
+                    .getForEntity(GET_QUOTES, QuoteDetailsResponse.class);
+            QuoteDetailsResponse body = response.getBody();
+            List<Quote> quotes = body.getQuotes();
+            List<QuotesHE> list = quotes.stream().map(mapper::modelToEntity)
+                    .collect(Collectors.toList());
+            quotesRepository.saveAll(list);
+        }
     }
 
+    @Transactional
     private void saveProducts(RestTemplate restTemplate) {
-        ResponseEntity<ProductDetailsResponse> response = restTemplate
-                .getForEntity(GET_PRODUCTS, ProductDetailsResponse.class);
-        ProductDetailsResponse body = response.getBody();
-        List<Product> quotes = body.getProducts();
-        List<ProductHE> list = quotes.stream().map(mapper::modelToEntity)
-                .collect(Collectors.toList());
-        productRepository.saveAll(list);
+        int size = productRepository.findAll().size();
+        if (size < 1 || size < 100) {
+            productRepository.deleteAll();
+            ResponseEntity<ProductDetailsResponse> response = restTemplate
+                    .getForEntity(GET_PRODUCTS, ProductDetailsResponse.class);
+            ProductDetailsResponse body = response.getBody();
+            List<Product> quotes = body.getProducts();
+            List<ProductHE> list = quotes.stream().map(mapper::modelToEntity)
+                    .collect(Collectors.toList());
+            productRepository.saveAll(list);
+        }
     }
 
     @PreDestroy
-    public void deleteRows() {
-        quotesRepository.deleteAll();
-        productRepository.deleteAll();
-    }
+    public void deleteAll() {}
 }
